Add files via upload
This is a pizza website. Having a GET PIZZA, MY ORDERS and SIGN IN Options. 
GET PIZZA::
So in the get pizza section you will meet variety of the pizzas images having the name and the text is on the image and also includes the prices on it.

MyOrders:
So in My Orders section you will see a beautiful receipt having current, date and time on it and the pizzas details and their prices also mentioned on it

Login:
In Login section, there is a login form having username, email and password detail on it